requirements.txt
python-dotenv
requests
icalendar
playwright
google-api-python-client
google-auth